backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
websockets==12.0
langchain-openai==0.1.7
langchain-core>=0.2.0,<0.3
openai>=1.40.0
sqlalchemy>=2.0.36
pydantic>=2.10.0,<3
python-dotenv==1.0.0
langdetect==1.0.9
requests==2.31.0
python-telegram-bot>=21.0
